[Resolved #661] - CRUD Car

+++ database/migrations/2016_08_16_034647_create_cars_table.php
@@ -16,7 +16,7 @@ class CreateCarsTable extends Migration
             $table->increments('id');
             $table->string('brand');
             $table->string('type');
-            $table->tinyInteger('year')->unsigned();
+            $table->smallInteger('year')->unsigned();
             $table->string('color');
             $table->string('plate')->unique()->index();
         });
